+++ src/region/index.js
@@ -1,17 +1,15 @@
 import wrapSetConfig from './wrapSetConfig';
 import wrapReducer from './wrapReducer';
-import wrapGet from './wrapGet';
 import wrapPrivate from './wrapPrivate';
 import wrapPublic from './wrapPublic';
-import wrapConnect from './wrapConnect';
+import wrapReact from './wrapReact';
 
 const compose = (...funcs) => funcs.reduce((a, b) => (...args) => a(b(...args)), arg => arg);
 
 export default compose(
-  wrapConnect,
+  wrapReact,
   wrapPublic,
   wrapPrivate,
-  wrapGet,
   wrapReducer,
   wrapSetConfig,
 )();