+++ src/region/wrapReact.js
@@ -7,7 +7,7 @@ import { isValidConnectKey } from '../util/isValidConnectKey';
 const Empty = () => null;
 
 export default (Region) => {
-  class RegionConnect extends Region {
+  class RegionReact extends Region {
     connectWith = (key, Display, option) => {
       const { connect } = this;
       return connect(key, option)(Display);
@@ -48,9 +48,9 @@ export default (Region) => {
       useEffect(() => {
         const unsubscribe = store.subscribe(() => setState(store.getState()));
         return () => unsubscribe();
-      });
+      }, []);
       return private_selectorFactory(key)(state);
     }
   }
-  return RegionConnect;
+  return RegionReact;
 };